feat: CommandAuth 분리, 유저 인증 여부 체크 함수 추가
- checkIsVerified 추가
- CommandAuth 함수 분리
